+++ dwds/lib/src/connections/app_connection.dart
@@ -19,7 +19,7 @@ class AppConnection {
 
   AppConnection(this.request, this._connection);
 
-  bool get isInKeepAlivePeriod => _connection.isInKeepAlivePeriod;
+  bool get isClosed => _connection.isClosed;
   void shutDown() => _connection.shutdown();
   bool get isStarted => _startedCompleter.isCompleted;
   Future<void> get onStart => _startedCompleter.future;